REFELECTION
in this activity, I have learned how to open the part in the PC with knowledge given by our lecturers. I have gained practical experience and deepened my understanding of the components that make up the CPU. Assembling and reassembling the CPU revealed the intricacies of each component, demanding precision and attention to detail. Before this, I just used the PC without any information about its components and the role that they played, now I know it and implement it in my future. Collaborating with classmates enriched the experience, highlighting the importance of effective communication in dealing with complex tasks.